+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 ### Fixed
 - Aggiornata la versione del broker per correzione su modulo di audit
 - Corretto bug in gestione del numero di protocollo (analisi del documentModel associato alla casella di posta)
-
+- Riconoscimento di file danneggiati all'interno di mail ricevute tramite interoperabilità: Viene inviata al mittente la notifica di eccezione.
 
 ## [3.0.7] - 2019-03-26